requirements.txt
openai==1.3.5
python-dotenv==1.0.0
rich==13.7.0